requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
streamlit>=1.22.0
ultralytics>=8.0.0
face-recognition
mediapipe
